pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "organvm-engine"
version = "0.1.0"
description = "Core governance, registry, and orchestration engine for the organvm eight-organ system"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
schema = ["jsonschema>=4.20"]
dev = ["pytest>=8.0", "jsonschema>=4.20", "ruff>=0.4", "pyright>=1.1"]

[project.scripts]
organvm = "organvm_engine.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "PTH", "RET", "SIM", "COM", "PL"]
ignore = [
    "E501",
    "E741",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLC0415",
    "PLR2004",
    "PLW0602",
    "PLW0603",
    "PLW2901",
]

[tool.ruff.lint.isort]
known-first-party = ["organvm_engine"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
